% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plik.R
\name{lin.regr}
\alias{lin.regr}
\title{Linear Regression}
\usage{
lin.regr(x, y, na.rm = FALSE)
}
\arguments{
\item{x}{numeric vector}

\item{y}{numeric vector}

\item{na.rm}{logical vector of length 1}
}
\value{
Returns a numeric vector with two elements,
the first one is slope of the linear function
which is best-fit for the data given, and the second
value is the y-intercept.
}
\description{
Function takes two vectors of the same length and
calculates the best-fit linear function using
Method of Least Squares.
}
