% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plik.R
\docType{package}
\name{Simregr-package}
\alias{Simregr}
\alias{Simregr-package}
\title{Simregr: Simple regressions and numpy functions}
\description{
There are implemented a few vectorised methods from numpy. And a linear regression.
}
\author{
\strong{Maintainer}: Bartosz Kitewski \email{01199235@pw.edu.pl}

Authors:
\itemize{
  \item Bartosz Kitewski \email{01199235@pw.edu.pl}
}

}
