% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plik.R
\name{lin.regr}
\alias{lin.regr}
\title{Linear Regression}
\usage{
lin.regr(x, y, na.rm = FALSE)
}
\arguments{
\item{x}{numeric vector}

\item{y}{numeric vector}

\item{na.rm}{logical vector of length 1}
}
\value{
Returns a numeric vector.
}
\description{
I've no time for this.
}
